chmux: add flush_interval configuration option
Forces flushing of transport sink at regular interval, even when
there is more data available for immediate sending.

+++ remoc/src/chmux/port_allocator.rs
@@ -19,7 +19,7 @@ struct PortAllocatorInner {
 
 impl PortAllocatorInner {
     fn is_available(&self) -> bool {
-        self.used.len() < self.limit as usize
+        self.used.len() < self.limit
     }
 
     fn try_allocate(&mut self, this: Arc<Mutex<PortAllocatorInner>>) -> Option<PortNumber> {
